change to nexInit to work on both 8 bit and 32 bit processors

+++ Nextion display/arduino_library_update/NexHardware.cpp
@@ -217,7 +217,7 @@ bool recvRetCommandFinished(uint32_t timeout)
 }
 
 
-bool nexInit(int Speed)
+bool nexInit(long Speed)
 {
     bool ret1 = false;
     bool ret2 = false;
